% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\alias{getBlockValues}
\alias{check_overlap}
\title{Get the detailed information of overlapped grids}
\usage{
overlap(r, geoms, return.id = FALSE)

getBlockValues(r, blocks, .progress = "none")

check_overlap(geoms, r)
}
\arguments{
\item{r}{Raster, RasterBrick, SpatialPixelsDataFrame or SpatialGridDataFrame object}

\item{geoms}{WKB objects, e.g.
\code{sf::st_as_binary(sf::st_geometry(basins), EWKB=TRUE)}}

\item{inheritParams}{st_extract}
}
\value{
\code{blocks}
\itemize{
\item \code{fraction}: fraction percentage
\item \code{area}: area in km^2
\item \code{row},\code{col}: begining \code{row} and \code{col} of the overlapped region
\item \code{nrow},\code{ncol}: number of the rows and columns of the overlapped region
}
}
\description{
Get the detailed information of overlapped grids

check_overlap
}
