% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_overlap.R
\name{rast_overlap}
\alias{rast_overlap}
\title{rast_overlap}
\usage{
rast_overlap(
  r,
  wkb,
  default_value = NA_real_,
  default_weight = NA_real_,
  include_xy = FALSE,
  include_cell = TRUE,
  include_area = FALSE,
  include_cols = NULL,
  new = TRUE,
  coverage_area = FALSE
)
}
\arguments{
\item{r}{raster}

\item{wkb}{wkb objects, e.g.
\code{sf::st_as_binary(sf::st_geometry(basins), EWKB=TRUE)}}
}
\value{
A data.frame, with the variables of
\itemize{
\item \code{values}            : values of raster \code{r}
\item \code{weight}            : cell area (in m^2)
\item \code{cell}              : cell id
\item \code{coverage_fraction} : fraction
}
}
\description{
Extract raster overlapping info by the function \verb{[sf.extract:::CPP_exact_extract()]}
}
\note{
!!!Please do not use function.
}
\keyword{internal}
