% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.extract-package.R
\docType{package}
\name{sf.extract-package}
\alias{sf.extract}
\alias{sf.extract-package}
\title{sf.extract: Fast Extraction from Raster Datasets using Polygons}
\description{
sf.extract for multiple raster files
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
