% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_overlap.R
\name{rast_overlap}
\alias{rast_overlap}
\title{rast_overlap}
\usage{
rast_overlap(
  r,
  wkb,
  default_value = NA_real_,
  default_weight = NA_real_,
  include_xy = FALSE,
  include_cell = TRUE,
  include_area = FALSE,
  include_cols = NULL,
  coverage_area = FALSE
)
}
\arguments{
\item{r}{raster}

\item{wkb}{WKB objects, e.g.
\code{sf::st_as_binary(sf::st_geometry(basins), EWKB=TRUE)}}
}
\description{
Extract raster overlapping info by the function \verb{[sf.extract:::CPP_exact_extract()]}
}
\keyword{internal}
